requirements.txt
python-telegram-bot<9.1
eyeD3<0.9
